refactor: replace getMonthName function to enumclass for connecting each month number with its name
